pyproject.toml
[build-system]
requires = ["setuptools>=65.0"]
build-backend = "setuptools.build_meta"

[project]
name = "study-materials-mcp"
version = "1.0.0"
description = "MCP server for managing and interacting with study materials"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
requires-python = ">=3.8"
dependencies = [
    "python-dotenv>=1.0.0",
    "PyPDF2>=3.0.1"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "pylint>=2.0.0"
]

[tool.setuptools.packages.find]
where = ["."]
